cluster/examples/requirements.txt
nats-py>=2.6.0
asyncio-nats-client>=0.11.5
